April release

+++ src/net/net_tcp.c
@@ -89,7 +89,7 @@ _Bool Net_SendStream(int32_t sock, const void *data, size_t len) {
 
 	ssize_t sent = 0;
 	while ((size_t) sent < buf.size) {
-		const ssize_t s = send(sock, (void *)(buf.data + sent), buf.size - sent, 0);
+		const ssize_t s = send(sock, (void *)(buf.data + sent), (int32_t) (buf.size - sent), 0);
 		if (s == -1) {
 			if (Net_GetError() != EWOULDBLOCK) {
 				Com_Warn("%s\n", Net_GetErrorString());
@@ -109,7 +109,7 @@ _Bool Net_ReceiveStream(int32_t sock, mem_buf_t *buf) {
 
 	buf->size = buf->read = 0;
 
-	const ssize_t received = recv(sock, (void *) buf->data, buf->max_size, 0);
+	const ssize_t received = recv(sock, (void *) buf->data, (int32_t) buf->max_size, 0);
 	if (received == -1) {
 
 		if (Net_GetError() == EWOULDBLOCK) {
